+++ src/App.js
@@ -1,10 +1,9 @@
-import style from './style/App.module.css';
-import Menu from './components/Menu';
+import Menu from './components/Menu/Menu';
 
 export default function App() {
-  return (
-    <div className={style.App}>
-      <Menu/>
-    </div>
-  );
+	return (
+		<div className='App'>
+			<Menu />
+		</div>
+	);
 }
