+++ src/components/Menu/menu.module.css
@@ -1,0 +1,121 @@
+.menuContainer {
+	position: relative;
+	margin-top: 20%;
+	perspective: 500px;
+}
+.menuButton {
+	width: 45px;
+	height: 45px;
+	background-color: rgb(117, 212, 232);
+	border: 2px solid #157c95;
+	-webkit-box-shadow: inset 0px 0px 0px 1.5px
+		rgba(255, 255, 255, 0.4);
+	box-shadow: inset 0px 0px 0px 1.5px
+		rgba(255, 255, 255, 0.4);
+	transform: rotateZ(45deg);
+	padding: 0;
+	filter: drop-shadow(
+		0px 5px 20px rgba(2, 113, 141, 0.5)
+	);
+}
+
+.menuButton svg {
+	transform: rotateZ(-45deg);
+	fill: #fbfbfc;
+	display: block;
+	width: 30px;
+	height: 30px;
+	margin: auto;
+}
+.menuWrapper {
+	opacity: 0;
+	position: relative;
+	top: -12px;
+	width: 500px;
+	height: 70px;
+	clip-path: polygon(
+		42% -2%,
+		50% 51%,
+		58% -2%,
+		100% 0,
+		100% 100%,
+		0 100%,
+		0 0
+	);
+	background: #abadb4;
+}
+.menuWrapper:after {
+	content: '';
+	position: absolute;
+	background: #e3e6e9;
+	top: 3.5px;
+	left: 3.4px;
+	right: 3.4px;
+	bottom: 3.5px;
+	clip-path: polygon(
+		41.9% 0%,
+		50% 58.2%,
+		58.1% 0,
+		100% 0,
+		100% 100%,
+		0 100%,
+		0 0
+	);
+}
+.menuWrapper:before {
+	content: '';
+	position: absolute;
+	background: #f9f9f9;
+	top: 2px;
+	left: 2px;
+	right: 2px;
+	bottom: 2px;
+	clip-path: polygon(
+		42% 0,
+		50% 55%,
+		58% 0,
+		100% 0,
+		100% 100%,
+		0 100%,
+		0 0
+	);
+}
+
+.menuList {
+	display: flex;
+	flex-direction: row;
+	z-index: 2;
+	position: relative;
+	list-style: none;
+	height: 100%;
+	margin: 0;
+	padding: 0;
+}
+.menuList li {
+	width: 100%;
+	justify-content: center;
+	align-items: center;
+	display: flex;
+	flex-direction: column;
+}
+.menuList li:nth-child(3) {
+	width: 70%;
+}
+
+.menuList li svg {
+	width: 1.6rem;
+}
+
+.menuList li p {
+	font-family: 'Dosis', sans-serif;
+	font-size: 12px;
+	color: #5d606f;
+	margin: 0;
+}
+
+.listButton {
+	width: 45%;
+	background-color: transparent;
+	border: none;
+	padding: 0;
+}
