+++ src/index.css
@@ -1,0 +1,21 @@
+body {
+	margin: 0;
+	font-family: -apple-system, BlinkMacSystemFont,
+		'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
+		'Cantarell', 'Fira Sans', 'Droid Sans',
+		'Helvetica Neue', sans-serif;
+	-webkit-font-smoothing: antialiased;
+	-moz-osx-font-smoothing: grayscale;
+	background-color: #f1f1f3;
+}
+
+code {
+	font-family: source-code-pro, Menlo, Monaco,
+		Consolas, 'Courier New', monospace;
+}
+
+.App {
+	text-align: center;
+	display: flex;
+	justify-content: center;
+}
