--- src/style/App.module.css
@@ -1,5 +1,0 @@
-.App {
-  text-align: center;
-  display: flex;
-  justify-content: center;
-}
